+++ packages/analytics/src/visitor/types.ts
@@ -1,4 +1,4 @@
-import type { UserProvidedData } from '../track/gtag';
+import type { UpdateVisitorDTO } from '../schema';
 import type { Environment, Platform, TrackTags } from '../track/types';
 
 export type VisitorProperties = Record<Lowercase<string>, string | number | boolean | null>;
@@ -9,13 +9,6 @@ export interface Visitor {
   platform: Platform;
   environment: Environment;
   tags: TrackTags;
-}
-
-export interface UpdateVisitorDTO {
-  user_id?: string;
-  distinct_id?: string;
-  data?: UserProvidedData;
-  properties?: VisitorProperties;
 }
 
 export type ThirdPartyUserSetter = (dto: UpdateVisitorDTO) => void;