+++ packages/utils/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @shware/utils
 
+## 1.5.1
+
+### Patch Changes
+
+- update visitor schema
+
 ## 1.5.0
 
 ### Minor Changes